modelo_deploy_azurefunction/requirements.txt
azure-functions
azure-storage-blob
tensorflow==2.15.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

